WoldVirtual_Crypto_3D/webgl threejs/WoldVirtualCrypto3D/requirements.txt
reflex==0.7.14
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
websockets==12.0
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.0
pycryptodome==3.15.0
web3==5.31.0
pytest==7.2.2
pytest-asyncio==0.20.3
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
sqlalchemy==2.0.25
alembic==1.13.1
ipfshttpclient==0.8.0
python-socketio==5.10.0